requirements.txt
requests>=2.32.0
PyYAML>=6.0
psycopg2-binary>=2.9.0